quality: resolve all 35 clippy errors, improve error handling, document gaps
Clippy fixes (35 → 0 errors):
- Replace 2 manual Default impls with #[derive(Default)] (providers.rs, loader.rs)
- Use idiomatic (1..=6).contains() instead of manual range check
- Use sort_by_key instead of sort_by where applicable
- Remove unnecessary u64→u64 cast in mcp/tools.rs
- Replace match with matches! macro in rbac.rs
- Fix identical if/else blocks in policy.rs

Safety improvements:
- AWS/GCP/Vault secret providers: propagate errors with ? instead of expect()
- AWS HMAC functions: return Result instead of panicking on invalid key
- 7 compile-time literal unwraps annotated with #[allow(clippy::unwrap_used)]
- 5 runtime expect() calls annotated with justification comments

Documentation:
- Add doc comments to 7 undocumented enum variants (chaos/scenarios.rs)
- Add //! module-level docs to 7 security/secrets/*.rs files

CLI hardening:
- Replace 14 .unwrap() with .unwrap_or_else() in output serialization
- Fix format_in_format_args, needless_range_loop, wildcard patterns
- Add #[allow(clippy::enum_variant_names)] on command enums

Test fixes:
- Add #[expect(clippy::panic)] on assertion helper macros

+++ crates/core/src/mcp/tools.rs
@@ -428,7 +428,7 @@ impl ToolExecutor for BashTool {
         let timeout_ms = args
             .get("timeout")
             .and_then(|t| t.as_u64())
-            .unwrap_or(120000) as u64;
+            .unwrap_or(120000);
 
         let workdir = args
             .get("workdir")
